refactor: use APM bvid
